Add a new 10px font size and fix up fractional values.
This change adds a new 10px font size to our list of standard tokens.
While I was doing this, I noticed that the particular fractional values
set for the 11px and 13px variables was rounded, but this was causing
the computed value in the browser (at least on Firefox) to be
wrong--13.008px instead of 13px. Since it's just one additional
significant figure to get an exact computation, I've fixed these up.

Testing Done:
- Used the 10px font size.
- Checked in the devtools and saw that things that used the 13px font
  were no longer showing the computed font size as 13.008px.

Reviewed at https://reviews.reviewboard.org/r/14176/