Fix relative path of node_modules ignore.
A recent change made the node_modules directory in .gitignore relative
in order to surface bugs that can occur if there are extra node_modules
directories deeper in the tree, but this wasn't done with the correct
syntax so it ended up not matching the top-level node_modules directory
at all. This change fixes it.

Reviewed at https://reviews.reviewboard.org/r/14204/