Better constrain dialog sizes with wide and tall content.
Dialogs are now better constrained on the screen, with maximum widths
for standard dialog sizes that keep it from reaching to the full width
of the viewport:

* Fit dialogs are now a minimum of 35vw (viewport width) and max of
  60vw, with a maximum height of 60vh (viewport height).

* Medium dialogs are now a minimum of 60vw and 60vh, and a maximum of
  70vw and 70vh.

* Large dialogs are now a minimum *and* maximum of 80vw and 80vh.

The minimum/maximum widths for Fit and Medium allow for some growth if
the content requires.

The dialog's body is also now set to show scrollbars if needed. This
avoids the scrollbar that would otherwise be placed on the `<dialog>`,
which covered the title and buttons area.

Testing Done:
Tested all the dialog sizes with minimum and maximum sizes, using
a variety of test content.

Reviewed at https://reviews.reviewboard.org/r/14648/