Add suppress checkboxes and better ARIA roles for dialogs.
It's common for some types of dialogs to have a suppression checkbox,
such as "Do not ask again." To ensure there's a standard approach for
this, dialogs can now have `canSuppress=true` set. This will place a
checkbox in the secondary actions area, at the far left.

The default text will be "Do not ask again," but this can be customized
by setting `suppressText=`.

It's up to the caller to determine how to suppress. The dialog doesn't
control any of this. It does provide the result of the checkbox though
as an `isSuppressed` property.

This will be used for upcoming confirmation utility dialogs.

Along with this, the ARIA role for modal dialogs has been updated to use
`alertdialog`. This will announce the dialog to screen readers, and is
the correct choice when popping up a modal dialog that requires
attention, rather than a utility dialog that is simply on screen.

Testing Done:
Tested with and without suppression checkboxes, and with custom
label text.

Reviewed at https://reviews.reviewboard.org/r/14649/