Add support for async click handlers for buttons.
A button's `onClick` can now be async. If a Promise is returned, the
button will be in a busy state until the Promise is resolved. This
ensures that the button can't be clicked a second time, and means the
caller doesn't have to micro-manage this.

Handlers that *do* need to micro-manage the state (such as setting the
busy state for multiple buttons at once) simply don't need to be async.
The button will work as before.

The Ink stories have been updated to show off click handlers for both
cases.

Testing Done:
Tested both async and standard `onClick` handlers.

Reviewed at https://reviews.reviewboard.org/r/14683/