Add IDs for the confirm and utility dialogs and their actions.
This allows callers to set an explicit ID when calling
`showConfirmDialog` or `showErrorDialog`. If not provided, a unique ID
will be computed.

That ID will also be the prefix for all actions. This makes it easier to
reference particular actions in these dialogs.

Testing Done:
Verified the IDs in both dialogs.

Made use of this in Review Board's unit tests to trigger actions in
dialogs.

Reviewed at https://reviews.reviewboard.org/r/14689/